Refactor completed. Went fairly smoothly. Tests passing and program works
