+++ message_handlers/message_handler_interface/message_handler_interface.go
@@ -1,4 +1,4 @@
-package message_handlers
+package message_handler_interface
 
 import (
 	"HigherLevelPerudoServer/messages"
